#130 Split test_ercot.py into unit test and integration tests.
This commit splits test_ercot.py into unit tests, which have no external
dependencies, and integration tests, which test pyiso's integration
with external systems (e.g. ERCOT's web servers).
