+++ tests/fixtures/ercot/real_time_system_conditions.html
@@ -1,0 +1,88 @@
+<html>
+<head>
+<script>
+  setTimeout("window.location.reload(true);",60000);
+</script>
+<title>Real-Time System Conditions</title>
+<meta http-equiv='pragma'  content='no-cache'>
+<link rel="stylesheet" type="text/css" href="/content/styles/cdr_reports.css" />
+<script>
+function open_window(url) {
+  window.open(url,"help",'toolbar=0,location=0,directories=0,status=0,menubar=0,scrollbars=yes,resizable=0,width=750,height=825');
+};
+</script>
+<!-- Google Analytics -->
+<script>
+(function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
+(i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
+m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
+})(window,document,'script','//www.google-analytics.com/analytics.js','ga');
+
+ga('create', 'UA-460876-10', 'auto');
+ga('send', 'pageview');
+</script>
+<!-- End Google Analytics -->
+</head>
+<body class="bodyStyle">
+<div style="display: inline-block;">
+<div>
+  <div class='header1'>Real-Time System Conditions</div>
+  <div class='helpLink'><a href="javascript:open_window('/help/popup/rtSysCondHelp')">Help?</a></div>
+</div>
+<br>
+<div class="schedTime rightAlign">Last Updated: Apr 14, 2016 18:38:40</div>
+<table class="tableStyle">
+<tbody>
+  <tr>
+    <td class="headerValueClass" colSpan="2">Frequency</td>
+  </tr>
+    <td class="tdLeft">Current Frequency</td>
+    <td class="labelClassCenter">59.998</td>
+  </tr>
+  </tr>
+    <td class="tdLeft">Instantaneous Time Error</td>
+    <td class="labelClassCenter">-17.468</td>
+  </tr>
+  <tr>
+    <td class="headerValueClass" colSpan="2">Real-Time Data</td>
+  </tr>
+  </tr>
+    <td class="tdLeft">Actual System Demand</td>
+    <td class="labelClassCenter">38850</td>
+  </tr>
+  </tr>
+    <td class="tdLeft">Total System Capacity (not including Ancillary Services)</td>
+    <td class="labelClassCenter">42514</td>
+  </tr>
+  </tr>
+    <td class="tdLeft">Total Wind Output</td>
+    <td class="labelClassCenter">5242</td>
+  </tr>
+  <tr>
+    <td class="headerValueClass" colSpan="2">DC Tie Flows</td>
+  </tr>
+    </tr>
+    <td class="tdLeft">DC_E (East)</td>
+    <td class="labelClassCenter">-31</td>
+  </tr>
+    </tr>
+    <td class="tdLeft">DC_L (Laredo VFT)</td>
+    <td class="labelClassCenter">0</td>
+  </tr>
+    </tr>
+    <td class="tdLeft">DC_N (North)</td>
+    <td class="labelClassCenter">0</td>
+  </tr>
+    </tr>
+    <td class="tdLeft">DC_R (Railroad)</td>
+    <td class="labelClassCenter">0</td>
+  </tr>
+    </tr>
+    <td class="tdLeft">DC_S (Eagle Pass)</td>
+    <td class="labelClassCenter">1</td>
+  </tr>
+  </tbody>
+</table>
+</div>
+</body>
+</html>
